requirements.txt
flask==2.2.2
discord.py==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==3.0.0
async-timeout==3.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
